Define `HitsCoordsSoA` header
Missing includes
